# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "2.0.0",
        "title": "Covid-19 API",
        "description": "REST API para registar pedidos de diagnostico no centro de análises",
        "contact": {
            "name": "Raquel Neto",
            "email": "8160129@estg.ipp.pt"
        }
    },
    "host": "http://localhost:3000",
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/tecnico/lancarPedido/:idutilizador": {
            "get": {
                "tags": [
                    "Tecnico"
                ],
                "summary": "O tecnico pode lançar o pedido através do ID.",
                "description": "O tecnico pode lançar o pedido através do ID no sistema.",
                "parameters": [{
                    "name": "user",
                    "in": "body",
                    "description": "Tecnico lança o pedido.",
                    "schema": {
                        "$ref": "#/definitions/User"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Tecnico lança o pedido",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/tecnico/aceitarPedido": {
            "get": {
                "tags": [
                    "Tecnico"
                ],
                "summary": "O tecnico pode aceitar o pedido.",
                "description": "O tecnico pode aceitar o pedido no sistema.",
                "parameters": [{
                    "name": "user",
                    "in": "body",
                    "description": "Tecnico aceita o pedido.",
                    "schema": {
                        "$ref": "#/definitions/User"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Tecnico aceita o pedido",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/tecnico/aceitarPedido/:idutilizador": {
            "put": {
                "tags": [
                    "Tecnico"
                ],
                "summary": "O tecnico pode aceitar o pedido através do ID.",
                "description": "O tecnico pode aceitar o pedido através do ID no sistema.",
                "parameters": [{
                    "name": "user",
                    "in": "body",
                    "description": "Tecnico aceitar o pedido.",
                    "schema": {
                        "$ref": "#/definitions/User"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Tecnico aceita o pedido",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/tecnico/eliminarPedido/:idutilizador": {
            "delete": {
                "tags": [
                    "Tecnico"
                ],
                "summary": "O tecnico pode eliminar o pedido através do ID.",
                "description": "O tecnico pode eliminar o pedido através do ID no sistema.",
                "parameters": [{
                    "name": "user",
                    "in": "body",
                    "description": "Tecnico elimina o pedido.",
                    "schema": {
                        "$ref": "#/definitions/User"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Tecnico elimina o pedido",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "required": [
                "name",
                "data_de_nascimento",
                "sexo",
                "numero_utente",
                "morada",
                "profissao",
                "telemovel",
                "email",
                "password",
                "date",
                "role",
                "Estado",
                "Teste_Realizado"

            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "data_de_nascimento": {
                    "type": "string"
                },
                "sexo": {
                    "type": "string"
                },
                "numero_utente": {
                    "type": "number",
                    "min": 9
                },
                "morada": {
                    "type": "string"
                },
                " profissao": {
                    "type": "string"
                },
                "telemovel": {
                    "type": "number",
                    "min": 9
                },
                "email": {
                    "type": "string",
                    "uniqueItems": true
                },
                "password": {
                    "type": "string",
                    "format": "password"
                },
                "date": {
                    "type": "string",
                    "format": "date"
                },
                "role": {
                    "type": "string",
                    "enum": ["PACIENTE", "ADMIN", "TECNICO"]
                },
                "Estado": {
                    "type": "integer",
                    "enum": ["Suspeito", "Infetado", "Não Infetado"],
                    "default": "Suspeito"
                },
                "Teste_Realizado": {
                    "type": "boolean"
                }
            }
        },
...
    }
}