# front-end/src/app/models/users.model.ts
export interface User {
    nome: String,
    data_de_nascimento: String,
    sexo: String,
    numero_utente: Number,
    morada: String,
    profissao: String,
    telemovel: Number,
    email: String,
    password:String,
    date: String,
    role: String,
    Estado: String,
    Teste_Realizado:Boolean

}